% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergwho.R
\name{dfmerg}
\alias{dfmerg}
\title{Merge the WHO growth velocity standard with your dataset}
\usage{
dfmerg(data, whodata, age, sex, interval)
}
\arguments{
\item{data}{input dataset}

\item{whodata}{input WHO standard data}

\item{age}{a numeric value in the data and whodata}

\item{sex}{a numeric value 1 for male and 2 for female}

\item{interval}{a numeric value indicating between measurement interval}
}
\value{
a tibble with merged dataset
}
\description{
This function merge the WHO growth velocity standard with your dataset.
}
\examples{
#First get the WHO growth velocity standard
data(package="dfvel")
#view the data that you want in the package (for example)
View(wholengthv09)
#merge data and whodata by age, sex and interval
#Make sure the variables used to merge the data set have similar names
dfmerg(data=data, whodata=wholengthv09, age=age, sex=sex, interval=interval)
}
