% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfvel-package.R
\docType{package}
\name{dfvel-package}
\alias{dfvel}
\alias{dfvel-package}
\title{dfvel: Compute Child Growth Velocity Using the WHO Standard}
\description{
The aim of the dfvel package is to simplify computation of child growth velocity in epidemiological studies. The package compute height, weight, and head circumference velocity from long form data using the numeric and LMS methods. The package uses the 2009 WHO growth velocity standard to compute individual growth velocity z score.
}
\author{
\strong{Maintainer}: Dawit Jember \email{devanhijember@gmail.com} (\href{https://orcid.org/0000-0003-3831-5465}{ORCID})

Authors:
\itemize{
  \item Dawit Jember \email{devanhijember@gmail.com}
}

}
\keyword{internal}
