% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heightv.R
\name{htvel}
\alias{htvel}
\title{Compute height velocity}
\usage{
htvel(data)
}
\arguments{
\item{input}{data that contains ID, time, age, height}
}
\value{
a tibble that contains ID, time, age, height, agediff, htdiff, hvel
}
\description{
This function compute height velocity using the numeric method in the long
form data
}
\examples{
##calculate height velocity
heightv<-htvel(sample_gdata)
}
