% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hghtvelz.R
\name{hvelzscore}
\alias{hvelzscore}
\title{Compute length velocity z score}
\usage{
hvelzscore(data)
}
\arguments{
\item{input}{data with ID, time, age, height, hl, hm, hs}
}
\value{
a tibble with individual z score
}
\description{
This function compute length velocity z score among children under
two years of age using the 2009 WHO growth velocity standard
}
\examples{
#example code
hvelzscore(data)
}
