% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_gdat}
\alias{sample_gdat}
\title{Description about sample_gdat dataset}
\format{
A data frame with 97 rows and 7 columns
\describe{
\item{ID}{A numeric value of individual identification}
\item{time}{A numeric value indicating measuremnt occassions}
\item{weight}{A numeric value of weight in kilogram}
\item{height}{A numeric value of height in centimeter}
\item{age}{A numeric value of age in months}
\item{sex}{A numeric value 1 represent male and 2 represent female}
\item{interval}{A numeric value indicating measuremnt intervals in months}
}
}
\usage{
sample_gdat
}
\description{
The 'sample_gdat' example dataset contains repeated length and weight measurements of 24 randomly
selected children under the age of two years. The data were obtained from a
community based cohort study conducted in Boricha and Bilate districts of Sidama Region, Ethiopia.
}
\keyword{datasets}
