% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lengthdat.R
\docType{data}
\name{wholengthv09}
\alias{wholengthv09}
\title{Description about 'wholengthv09' dataset}
\format{
A data frame with 170 rows and 13 columns
\describe{
\item{age}{A numeric value of age in months}
\item{interval}{A numeric value indicating interval for length velocity}
\item{hl}{A numeric value of lambda}
\item{hm}{A numeric value of median length velocity}
\item{hs}{A numeric value of coefficient of variation}
\item{minus3sd}{Minus three standard deviation from the median}
\item{minus2sd}{Minus two standard deviation from the median}
\item{minus1sd}{Minus one standard deviation from the median}
\item{median}{A numeric value of median lenght velocity}
\item{plus1sd}{Plus one standard deviation from the median length velocity}
\item{plus2sd}{Plus two standard deviation from the median length velocity}
\item{plus3sd}{Plus two standard deviation from the median length velocity}
\item{sex}{A numeric value 1 represent male and 2 represent female}
}
}
\source{
World Health Organization (WHO). Child growth standards for length velocity.
Available at \url{https://www.who.int/tools/child-growth-standards/standards/length-velocity}
}
\usage{
wholengthv09
}
\description{
The 'wholengthv09' dataset contains the 2009 WHO length velocity standard
for children under two years of age.
}
\keyword{datasets}
