% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightv.R
\name{wtvel}
\alias{wtvel}
\title{Compute weight velocity}
\usage{
wtvel(data)
}
\arguments{
\item{input}{data contains ID, time, age, weight}
}
\value{
a tibble that contains ID, time, age, weight, agediff, wtdiff, wvel
}
\description{
This function compute weight velocity using the numeric method in the
long form data
}
\examples{
#Compute weight velocity
weightv<-wtvel(data)
}
