% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wghtvelz.R
\name{wvelzscore}
\alias{wvelzscore}
\title{Compute weight velocity z score}
\usage{
wvelzscore(data)
}
\arguments{
\item{data}{}
}
\value{
weight velocity z score
}
\description{
This function compute weight velocity Z score based on the WHO standard
}
\examples{
#Get individual weight velocity z score
wvelzscore(data)
}
